+++ HISTORY.rst
@@ -9,7 +9,7 @@ Unreleased
 **Data Migration**
 
 * ALCN-1224: set fiscal position to "Wholesaler without APB" to customer in category "Grossistes vétérinaires et Callcenter"
-
+* ALCN-1233: Set fiscal position on suppliers
 
 **Features and Improvements**